pyproject.toml
[tool.poetry]
name = "chatbot-backend"
version = "0.1.0"
description = ""
authors = ["akshatj2209 <akshatj2209@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
pymongo = {extras = ["srv"], version = "^4.3.3"}
motor = "^3.1.1"
pydantic = {extras = ["email"], version = "^1.8.2"}
python-dotenv = "^0.19.0"
google-generativeai = "^0.7.2"
sqlalchemy = "^2.0.34"

[tool.poetry.dev-dependencies]
pytest = "^6.2"

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn chatbot_backend.main:app --reload"
